Rename {assumed_size,deprecated}_character_syntax.rs
