+++ fortitude/src/rules/obsolescent/deprecated_character_syntax.rs
